refactor: fixture 위치 변경
