+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/acceptance/post/PostRepositorySearchAcceptanceTest.java
@@ -1,7 +1,7 @@
 package com.woowacourse.pickgit.acceptance.post;
 
-import static com.woowacourse.pickgit.query.fixture.TUser.GUEST;
-import static com.woowacourse.pickgit.query.fixture.TUser.NEOZAL;
+import static com.woowacourse.pickgit.common.fixture.TUser.GUEST;
+import static com.woowacourse.pickgit.common.fixture.TUser.NEOZAL;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.woowacourse.pickgit.acceptance.AcceptanceTest;
