+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/common/fixture/Act.java
@@ -1,4 +1,4 @@
-package com.woowacourse.pickgit.query.fixture;
+package com.woowacourse.pickgit.common.fixture;
 
 import static io.restassured.RestAssured.given;
