+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/common/fixture/TContact.java
@@ -1,4 +1,4 @@
-package com.woowacourse.pickgit.query.fixture;
+package com.woowacourse.pickgit.common.fixture;
 
 import com.woowacourse.pickgit.portfolio.presentation.dto.request.ContactRequest;
 import com.woowacourse.pickgit.portfolio.presentation.dto.response.ContactResponse;
