+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/common/fixture/UnLoginAndThenAct.java
@@ -1,4 +1,4 @@
-package com.woowacourse.pickgit.query.fixture;
+package com.woowacourse.pickgit.common.fixture;
 
 import static org.assertj.core.api.Assertions.assertThat;
