+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/common/mockapi/MockTagApiRequester.java
@@ -3,7 +3,7 @@ package com.woowacourse.pickgit.common.mockapi;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.woowacourse.pickgit.exception.platform.PlatformHttpErrorException;
-import com.woowacourse.pickgit.query.fixture.TRepository;
+import com.woowacourse.pickgit.common.fixture.TRepository;
 import com.woowacourse.pickgit.tag.infrastructure.PlatformTagApiRequester;
 import java.util.Map;
