+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/query/acceptance/post/PostAcceptanceTest_Query.java
@@ -1,15 +1,15 @@
 package com.woowacourse.pickgit.query.acceptance.post;
 
-import static com.woowacourse.pickgit.query.fixture.TPost.KEVINPOST;
-import static com.woowacourse.pickgit.query.fixture.TPost.MARKPOST;
-import static com.woowacourse.pickgit.query.fixture.TPost.NEOZALPOST;
-import static com.woowacourse.pickgit.query.fixture.TUser.DANI;
-import static com.woowacourse.pickgit.query.fixture.TUser.GUEST;
-import static com.woowacourse.pickgit.query.fixture.TUser.KEVIN;
-import static com.woowacourse.pickgit.query.fixture.TUser.KODA;
-import static com.woowacourse.pickgit.query.fixture.TUser.MARK;
-import static com.woowacourse.pickgit.query.fixture.TUser.NEOZAL;
-import static com.woowacourse.pickgit.query.fixture.TUser.모든유저;
+import static com.woowacourse.pickgit.common.fixture.TPost.KEVINPOST;
+import static com.woowacourse.pickgit.common.fixture.TPost.MARKPOST;
+import static com.woowacourse.pickgit.common.fixture.TPost.NEOZALPOST;
+import static com.woowacourse.pickgit.common.fixture.TUser.DANI;
+import static com.woowacourse.pickgit.common.fixture.TUser.GUEST;
+import static com.woowacourse.pickgit.common.fixture.TUser.KEVIN;
+import static com.woowacourse.pickgit.common.fixture.TUser.KODA;
+import static com.woowacourse.pickgit.common.fixture.TUser.MARK;
+import static com.woowacourse.pickgit.common.fixture.TUser.NEOZAL;
+import static com.woowacourse.pickgit.common.fixture.TUser.모든유저;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.woowacourse.pickgit.acceptance.AcceptanceTest;
