+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/unit/tag/infrastructure/GithubTagExtractorTest.java
@@ -1,6 +1,6 @@
 package com.woowacourse.pickgit.unit.tag.infrastructure;
 
-import static com.woowacourse.pickgit.query.fixture.TRepository.PICK_GIT;
+import static com.woowacourse.pickgit.common.fixture.TRepository.PICK_GIT;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
